eval handler with remote capability; release

+++ server/executionhandlers/eval/execution.rb
@@ -35,7 +35,7 @@ module CPEE
         hw = CPEE::Persistence::extract_item(id,opts,'executionhandler')
         endpoints = CPEE::Persistence::extract_list(id,opts,'endpoints')
         dataelements = CPEE::Persistence::extract_list(id,opts,'dataelements')
-        attributes = CPEE::Persistence::extract_list(id,opts,'attributes')
+        attributes = CPEE::Persistence::extract_list(id,opts,'attributes').to_h
         positions = CPEE::Persistence::extract_set(id,opts,'positions')
         positions.map! do |k, v|
           [ k, v, CPEE::Persistence::extract_item(id,opts,File.join('positions',k,'@passthrough')) ]
@@ -46,9 +46,14 @@ module CPEE
         iopts[:redis_url] = opts[:redis_url]
         iopts[:redis_db] = opts[:redis_db]
         iopts[:workers] = opts[:workers]
-        iopts[:global_executionhandlers] = opts[:global_executionhandlers]
-        iopts[:executionhandlers] = opts[:executionhandlers]
         iopts[:executionhandler] = hw
+        if attributes.has_key?('remote')
+          uri = URI::parse(attributes['remote'])
+          iopts[:executionhandlers] = File.join(uri.path,File.basename(opts[:executionhandlers]))
+        else
+          iopts[:executionhandlers] = opts[:executionhandlers]
+          iopts[:global_executionhandlers] = opts[:global_executionhandlers]
+        end
 
         File.open(File.join(opts[:instances],id.to_s,File.basename(ExecutionHandler::Eval::BACKEND_OPTS)),'w') do |f|
           YAML::dump(iopts,f)
@@ -56,24 +61,55 @@ module CPEE
         template = ERB.new(File.read(ExecutionHandler::Eval::BACKEND_TEMPLATE), trim_mode: '-')
         res = template.result_with_hash(dsl: dsl, dataelements: dataelements, endpoints: endpoints, positions: positions)
         File.write(File.join(opts[:instances],id.to_s,ExecutionHandler::Eval::BACKEND_INSTANCE),res)
+        if attributes.has_key?('remote')
+          uri = URI::parse(attributes['remote'])
+          Net::SSH.start(uri.host,uri.user,:keys => [ opts[:ssh_key] ] ) do |ssh|
+            ssh.exec!("rm -rf #{File.join(uri.path,id.to_s,'*')}")
+            ssh.scp.upload!(File.join(opts[:instances],id.to_s),uri.path,:recursive=>true)
+          end
+          File.write(File.join(opts[:instances],id.to_s,'.remote'),attributes['remote'])
+        end
       end
 
       def self::run(id,opts)
-        exe = File.join(opts[:instances],id.to_s,File.basename(ExecutionHandler::Eval::BACKEND_RUN))
-        pid = Kernel.spawn(opts[:libs_preloaderrun] + ' ' + exe , :pgroup => true, :in => '/dev/null', :out => exe + '.out', :err => exe + '.err')
-        Process.detach pid
-        File.write(exe + '.pid',pid)
+        if File.exist? File.join(opts[:instances],id.to_s,'.remote')
+          uri = URI::parse(File.read(File.join(opts[:instances],id.to_s,'.remote')))
+          exe = File.join(uri.path,id.to_s,File.basename(BACKEND_RUN))
+          Net::SSH.start(uri.host,uri.user,:keys => [ opts[:ssh_key] ] ) do |ssh|
+            ssh.exec!("ruby #{exe} >#{exe}.out 2>#{exe}.err &")
+          end
+        else
+          exe = File.join(opts[:instances],id.to_s,File.basename(ExecutionHandler::Eval::BACKEND_RUN))
+          pid = Kernel.spawn(opts[:libs_preloaderrun] + ' ' + exe , :pgroup => true, :in => '/dev/null', :out => exe + '.out', :err => exe + '.err')
+          Process.detach pid
+          File.write(exe + '.pid',pid)
+        end
       end
 
       def self::stop(id,opts) ### return: bool to tell if manually changing redis is necessary
-        exe = File.join(opts[:instances],id.to_s,File.basename(ExecutionHandler::Eval::BACKEND_RUN))
-        pid = File.read(exe + '.pid') rescue nil
-        if pid && (Process.kill(0, pid.to_i) rescue false)
-          Process.kill('HUP', pid.to_i) rescue nil
-          false
-        else # its not running, so clean up
-          File.unlink(exe + '.pid') rescue nil
-          true
+        if File.exist? File.join(opts[:instances],id.to_s,'.remote')
+          uri = URI::parse(File.read(File.join(opts[:instances],id.to_s,'.remote')))
+          exe = File.join(uri.path,id.to_s,File.basename(BACKEND_RUN))
+          Net::SSH.start(uri.host,uri.user,:keys => [ opts[:ssh_key] ] ) do |ssh|
+            pid = ssh.exec!("cat #{exe}.pid 2>/dev/null")
+            if pid != '' && ssh.exec!("kill -0 #{pid} >/dev/null 2>&1; echo $?").strip == '0'
+              ssh.exec!("kill -SIGHUP #{pid}")
+              false
+            else
+              ssh.exec!("rm #{exe}.pid")
+              true
+            end
+          end
+        else
+          exe = File.join(opts[:instances],id.to_s,File.basename(ExecutionHandler::Eval::BACKEND_RUN))
+          pid = File.read(exe + '.pid') rescue nil
+          if pid && (Process.kill(0, pid.to_i) rescue false)
+            Process.kill('HUP', pid.to_i) rescue nil
+            false
+          else # its not running, so clean up
+            File.unlink(exe + '.pid') rescue nil
+            true
+          end
         end
       end
     end
